Flytter valideringsfunksjoner til egen fil
