+++ src/main/kotlin/no/skatteetaten/aurora/boober/feature/ToxiproxySidecarSpec.kt
@@ -11,13 +11,12 @@ import no.skatteetaten.aurora.boober.service.UserDetailsProvider
 import no.skatteetaten.aurora.boober.service.resourceprovisioning.DatabaseSchemaProvisioner
 import no.skatteetaten.aurora.boober.utils.UrlParser
 import no.skatteetaten.aurora.boober.utils.toMap
-import no.skatteetaten.aurora.boober.utils.whenTrue
 
 private const val FEATURE_NAME = "toxiproxy"
 
 private const val FIRST_PORT_NUMBER = 18000 // The first Toxiproxy port will be set to this number
 
-private const val MAIN_PROXY_NAME = "app"
+internal const val MAIN_PROXY_NAME = "app"
 
 internal object ToxiproxyField {
     const val version = "$FEATURE_NAME/version"
@@ -31,26 +30,26 @@ data class ToxiproxyConfig(
     val enabled: Boolean = true
 )
 
-private data class ToxiproxyEndpoint(
+internal data class ToxiproxyEndpoint(
     val proxyName: String,
     val varName: String,
     val enabled: Boolean
 )
 
-private data class ToxiproxyServerAndPortVars(
+internal data class ToxiproxyServerAndPortVars(
     val proxyname: String,
     val serverVar: String?,
     val portVar: String?,
     val enabled: Boolean
 )
 
-private data class ToxiproxyDatabase(
+internal data class ToxiproxyDatabase(
     val proxyName: String,
     val databaseName: String,
     val enabled: Boolean
 )
 
-private data class ToxiproxyProxy(
+internal data class ToxiproxyProxy(
     val proxyName: String,
     val enabled: Boolean,
     val initialEnabledState: Boolean,
@@ -122,123 +121,6 @@ private data class ToxiproxyProxy(
 
 internal val AuroraDeploymentSpec.toxiproxyVersion: String?
     get() = featureEnabled(FEATURE_NAME) { this[ToxiproxyField.version] }
-
-internal fun AuroraDeploymentSpec.validateToxiproxy(): List<AuroraDeploymentSpecValidationException> =
-    listOf(
-        invalidCombinationOfToxiproxyFieldsErrors(),
-        missingOrInvalidEndpointVariableErrors(),
-        missingServerAndPortVariableErrors(),
-        missingNamedDbErrors(),
-        invalidDbConfigErrors(),
-        proxyNameError(),
-        duplicateUrlVariableErrors(),
-        duplicateServerAndPortVariableErrors(),
-        duplicateDatabaseNameErrors()
-    ).flatten()
-
-private fun AuroraDeploymentSpec.invalidCombinationOfToxiproxyFieldsErrors() =
-    extractEnabledToxiproxyProxies().mapNotNull { it.invalidCombinationError() }
-
-// Validate that the variables given in toxiproxy/proxies/<proxy name>/urlVariable exist
-private fun AuroraDeploymentSpec.missingOrInvalidEndpointVariableErrors() =
-    getToxiproxyEndpointEnvVars().mapNotNull { varName ->
-        val envVar = getOrNull<String>("config/$varName")
-        if (envVar == null) {
-            "Found Toxiproxy config for endpoint named $varName, but there is no such environment variable."
-        } else if (!UrlParser(envVar).isValid()) {
-            "The format of the URL \"$envVar\" given by the config variable $varName is not supported."
-        } else null
-    }.map(::AuroraDeploymentSpecValidationException)
-
-// Validate that the variables given in toxiproxy/proxies/<proxy name>/serverVariable
-// and toxiproxy/proxies/<proxy name>/portVariable exist
-private fun AuroraDeploymentSpec.missingServerAndPortVariableErrors() =
-    extractToxiproxyServersAndPorts().flatMap { serverAndPortsVars ->
-
-        val (proxyName, serverVar, portVar) = serverAndPortsVars
-        val serverAndPort = listOf("server" to serverVar, "port" to portVar)
-
-        serverAndPort.mapNotNull { (name, value) ->
-            if (value.isNullOrBlank()) {
-                return@mapNotNull "The $name variable is missing for the Toxiproxy proxy named $proxyName."
-            }
-            val isServerOrPortMissingFromConfig = this.getSubKeyValues("config").none { it == value }
-            if (isServerOrPortMissingFromConfig) {
-                "Found Toxiproxy config for a $name variable named $value, " +
-                    "but there is no such environment variable."
-            } else null
-        }.map(::AuroraDeploymentSpecValidationException)
-    }
-
-// Validate that the "database" and "databaseName" properties are used correctly
-private fun AuroraDeploymentSpec.invalidDbConfigErrors(): List<AuroraDeploymentSpecValidationException> {
-    val errors = mutableListOf<String>()
-    if (isSimplifiedAndEnabled("database") && extractToxiproxyNamedDatabases().isNotEmpty()) {
-        errors.add(
-            "Found named database(s) in the Toxiproxy config, although the database config is simplified. " +
-                "Did you mean to use the property \"database\" instead of \"databaseName\"?"
-        )
-    }
-    val numberOfDefaultDbConfigs = extractEnabledToxiproxyProxies().count { it.isDefaultDatabaseProxy() }
-    if (numberOfDefaultDbConfigs > 1) {
-        errors.add("The \"database\" property may only be used once in the Toxiproxy config.")
-    }
-    if (numberOfDefaultDbConfigs > 0 && !isSimplifiedConfig("database")) {
-        errors.add(
-            "It is not possible to set up a Toxiproxy proxy with the \"database\" property when the database " +
-                "config is not simplified. Did you mean to use \"databaseName\"?"
-        )
-    }
-    if (numberOfDefaultDbConfigs > 0 && isSimplifiedAndDisabled("database")) {
-        errors.add("It is not possible to set up a Toxiproxy proxy for a disabled database.")
-    }
-    return errors.map(::AuroraDeploymentSpecValidationException)
-}
-
-// Validate that for every database in toxiproxy/proxies/<proxy name>/databaseName,
-// there is a corresponding database in the spec
-private fun AuroraDeploymentSpec.missingNamedDbErrors() =
-    if (!isSimplifiedAndEnabled("database"))
-        extractToxiproxyNamedDatabases()
-            .map { it.databaseName }
-            .mapNotNull {
-                val dbNames = getSubKeyValues("database")
-                if (!dbNames.contains(it)) {
-                    AuroraDeploymentSpecValidationException(
-                        "Found Toxiproxy config for database named $it, " +
-                            "but there is no such database configured."
-                    )
-                } else null
-            }
-    else emptyList()
-
-// Validate that the proxyname used for incoming calls is not duplicated
-private fun AuroraDeploymentSpec.proxyNameError() = extractEnabledToxiproxyProxies()
-    .any { it.proxyName == MAIN_PROXY_NAME }
-    .whenTrue { listOf(AuroraDeploymentSpecValidationException("The name \"$MAIN_PROXY_NAME\" is reserved for the proxy for incoming calls.")) }
-    ?: emptyList()
-
-// Validate that all urlVariable values are unique
-private fun AuroraDeploymentSpec.duplicateUrlVariableErrors() = extractToxiproxyEndpoints()
-    .groupingBy { it.varName }
-    .eachCount()
-    .filter { it.value > 1 }
-    .map { AuroraDeploymentSpecValidationException("The url variable \"${it.key}\" is referred to by several proxies.") }
-
-// Validate that all serverVariable and portVariable combinations are unique
-private fun AuroraDeploymentSpec.duplicateServerAndPortVariableErrors() = extractToxiproxyServersAndPorts()
-    .groupingBy { Pair(it.serverVar, it.portVar) }
-    .eachCount()
-    .filter { it.value > 1 }
-    .map { "The server and port variables \"${it.key.first}\" and \"${it.key.second}\" are referred to by several proxies." }
-    .map(::AuroraDeploymentSpecValidationException)
-
-// Validate that all databaseName values are unique
-private fun AuroraDeploymentSpec.duplicateDatabaseNameErrors() = extractToxiproxyDatabases()
-    .groupingBy { it.databaseName }
-    .eachCount()
-    .filter { it.value > 1 }
-    .map { AuroraDeploymentSpecValidationException("The database name \"${it.key}\" is referred to by several proxies.") }
 
 private fun AuroraDeploymentSpec.endpointsFromConfig(initialPort: Int) =
     extractToxiproxyEndpoints().mapIndexedNotNull { i, (proxyname, varname, enabled) ->
@@ -356,7 +238,7 @@ internal fun List<Container>.overrideEnvVarsWithProxies(adc: AuroraDeploymentSpe
 internal fun List<ToxiproxyConfig>.getNextPortNumber(numberIfEmpty: Int) =
     maxOfOrNull { it.listen.substringAfter(':').toInt() }.let { if (it == null) numberIfEmpty else it + 1 }
 
-private fun AuroraDeploymentSpec.extractEnabledToxiproxyProxies() = getSubKeyValues(ToxiproxyField.proxies).map {
+internal fun AuroraDeploymentSpec.extractEnabledToxiproxyProxies() = getSubKeyValues(ToxiproxyField.proxies).map {
     ToxiproxyProxy(
         proxyName = it,
         enabled = get("${ToxiproxyField.proxies}/$it/enabled"),
@@ -370,16 +252,16 @@ private fun AuroraDeploymentSpec.extractEnabledToxiproxyProxies() = getSubKeyVal
 }.filter { it.enabled }
 
 // Return a list of proxynames and corresponding environment variable names
-private fun AuroraDeploymentSpec.extractToxiproxyEndpoints() =
+internal fun AuroraDeploymentSpec.extractToxiproxyEndpoints() =
     extractEnabledToxiproxyProxies().mapNotNull { it.toToxiproxyEndpoint() }
 
-private fun AuroraDeploymentSpec.extractToxiproxyServersAndPorts() =
+internal fun AuroraDeploymentSpec.extractToxiproxyServersAndPorts() =
     extractEnabledToxiproxyProxies().mapNotNull { it.toToxiproxyServerAndPortVars() }
 
-private fun AuroraDeploymentSpec.extractToxiproxyDatabases() =
+internal fun AuroraDeploymentSpec.extractToxiproxyDatabases() =
     extractEnabledToxiproxyProxies().mapNotNull { it.toToxiproxyDatabase(this["$databaseDefaultsKey/name"]) }
 
-private fun AuroraDeploymentSpec.extractToxiproxyNamedDatabases() = extractEnabledToxiproxyProxies()
+internal fun AuroraDeploymentSpec.extractToxiproxyNamedDatabases() = extractEnabledToxiproxyProxies()
     .filter { it.isNamedDatabaseProxy() }
     .mapNotNull { it.toToxiproxyDatabase() }
 
@@ -392,5 +274,5 @@ private fun String.convertToProxyUrl(port: Int): String =
         .withModifiedPort(port)
         .makeString()
 
-private fun AuroraDeploymentSpec.getToxiproxyEndpointEnvVars(): List<String> =
+internal fun AuroraDeploymentSpec.getToxiproxyEndpointEnvVars(): List<String> =
     extractToxiproxyEndpoints().map { it.varName }
